Problem: todolist example doesn't expose the evented nature
Judging just by the exposed functionality, it is impossible
to tell that todolist is using an immutable database such as
PumpkinDB.

Solution: exposes an 'event log' function to the users

This changes the way todo item events are collected (additional
journalling is now done to simplify the recovery of the complete
event log). The journal looks like this:

`[uuid][HLC] => [eventkey]`

This way we can simply iterate over all keys that start with a
given `uuid`.

This changes still continues the "tradition" of a terribly
organized code. Therefore I invite anybody, including future myself,
to get in and reorganize it, while documenting it. I just wanted
to get some basic functonality out.